requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
Flask==3.1.3
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.11
MarkupSafe==3.0.3
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
py4j==0.10.9.9
pycurl==7.45.7
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
warc==0.2.1
warcio==1.7.5
Werkzeug==3.1.6
